+++ Tests/Public/Core/Row/MetalRowTests.swift
@@ -177,14 +177,12 @@ class MetalRowTests: GRDBTestCase {
         }
     }
     
-    func testRowSubscriptIsCaseInsensitive() {
+    func testRowIsCaseInsensitive() {
         assertNoError {
             let dbQueue = DatabaseQueue()
-            try dbQueue.inDatabase { db in
-                try db.execute("CREATE TABLE stuffs (name TEXT)")
-                try db.execute("INSERT INTO stuffs (name) VALUES ('foo')")
+            dbQueue.inDatabase { db in
                 var rowFetched = false
-                for row in Row.fetch(db, "SELECT nAmE FROM stuffs") {
+                for row in Row.fetch(db, "SELECT 'foo' AS nAmE") {
                     rowFetched = true
                     XCTAssertEqual(row["name"], "foo".databaseValue)
                     XCTAssertEqual(row["NAME"], "foo".databaseValue)
@@ -198,14 +196,31 @@ class MetalRowTests: GRDBTestCase {
         }
     }
     
+    func testRowIsCaseInsensitiveAndReturnsLeftmostMatchingColumn() {
+        assertNoError {
+            let dbQueue = DatabaseQueue()
+            dbQueue.inDatabase { db in
+                var rowFetched = false
+                for row in Row.fetch(db, "SELECT 1 AS name, 2 AS NAME") {
+                    rowFetched = true
+                    XCTAssertEqual(row["name"], 1.databaseValue)
+                    XCTAssertEqual(row["NAME"], 1.databaseValue)
+                    XCTAssertEqual(row["NaMe"], 1.databaseValue)
+                    XCTAssertEqual(row.value(named: "name") as Int, 1)
+                    XCTAssertEqual(row.value(named: "NAME") as Int, 1)
+                    XCTAssertEqual(row.value(named: "NaMe") as Int, 1)
+                }
+                XCTAssertTrue(rowFetched)
+            }
+        }
+    }
+    
     func testRowHasColumnIsCaseInsensitive() {
         assertNoError {
             let dbQueue = DatabaseQueue()
-            try dbQueue.inDatabase { db in
-                try db.execute("CREATE TABLE stuffs (name TEXT)")
-                try db.execute("INSERT INTO stuffs (name) VALUES ('foo')")
+            dbQueue.inDatabase { db in
                 var rowFetched = false
-                for row in Row.fetch(db, "SELECT nAmE, 1 AS foo FROM stuffs") {
+                for row in Row.fetch(db, "SELECT 'foo' AS nAmE, 1 AS foo") {
                     rowFetched = true
                     XCTAssertTrue(row.hasColumn("name"))
                     XCTAssertTrue(row.hasColumn("NAME"))